Refactor: Move 'depends_openai_client_chat_completion_request' to openai_clients.py
- This function now resides in a more appropriate location, improving code organization.
- Removed unused imports from chat.py and added necessary imports to openai_clients.py to support the relocated function.
