+++ languru/server/deps/openai_clients.py
@@ -1,13 +1,15 @@
 import time
 from logging import Logger
-from typing import Any, Dict, List, Optional, Sequence, Text, Union
+from typing import Any, Dict, List, Optional, Sequence, Text, Tuple, Union
 
-from fastapi import Query, Request
+from fastapi import Body, Depends, Query, Request
+from fastapi.exceptions import HTTPException
 from openai import AzureOpenAI, OpenAI, OpenAIError
 from openai.types import Model
 from pydantic import BaseModel
 
 from languru.config import logger as languru_logger
+from languru.examples.openapi.chat import chat_openapi_examples
 from languru.exceptions import (
     CredentialsNotProvided,
     ModelNotFound,
@@ -19,7 +21,8 @@ from languru.openai_plugins.clients.groq import GroqOpenAI
 from languru.openai_plugins.clients.pplx import PerplexityOpenAI
 from languru.openai_plugins.clients.voyage import VoyageOpenAI
 from languru.server.config import APP_STATE_SETTINGS
-from languru.server.utils.common import get_value_from_app
+from languru.server.utils.common import get_value_from_app, to_openapi_examples
+from languru.types.chat.completions import ChatCompletionRequest
 from languru.types.models import (
     MODELS_ANTHROPIC,
     MODELS_AZURE_OPENAI,
@@ -30,6 +33,7 @@ from languru.types.models import (
     MODELS_VOYAGE,
 )
 from languru.types.organizations import OrganizationType, to_org_type
+from languru.utils.common import display_object
 
 
 class OpenaiModels:
@@ -383,3 +387,32 @@ class OpenaiClients(OpenaiModels, OpenaiDepends):
 
 
 openai_clients = OpenaiClients()
+
+
+def depends_openai_client_chat_completion_request(
+    request: "Request",
+    org_type: Optional[OrganizationType] = Depends(openai_clients.depends_org_type),
+    chat_completion_request: ChatCompletionRequest = Body(
+        ...,
+        openapi_examples=to_openapi_examples(chat_openapi_examples),
+    ),
+) -> Tuple[OpenAI, ChatCompletionRequest]:
+    logger = get_value_from_app(
+        request.app, key="logger", value_typing=Logger, default=languru_logger
+    )
+
+    if org_type is None:
+        org_type = openai_clients.org_from_model(chat_completion_request.model)
+    if org_type is None:
+        raise HTTPException(status_code=400, detail="Organization type not found.")
+
+    chat_completion_request.model = openai_clients.model_strip_org(
+        chat_completion_request.model, org_type
+    )
+    openai_client = openai_clients.org_to_openai_client(org_type)
+    logger.debug(
+        f"Organization type: '{org_type}', "
+        + f"openAI client: '{display_object(openai_client)}', "
+        + f"model: '{chat_completion_request.model}'"
+    )
+    return (openai_client, chat_completion_request)
